+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./video/index.ts";
 export * from "./audio/index.ts";
 export * from "./files.ts";
+export * from "./canvas.ts";